separate NoteListItemText into own component
as it's own component, it can fetch contents for note individually.
So we'll have it fetch a note from the backend. BUT -- we'll already
have that note in the cache, since we fetched all notes for the
current folder. So even though it looks like each component will
fetch a note from the backend, it'll only fetch from the cache
